Refactor index page
